# Optimize `Dictionary<>.TryInsert()` by using `ref var`