Benchmarking shows some improvement:
Windows, x64, .NET9
```
| Method                          | Mean     | Error     | StdDev    | Ratio | RatioSD |
|-------------------------------- |---------:|----------:|----------:|------:|--------:|
| 'Old Dictionary this[i] setter' | 5.170 ns | 0.2647 ns | 0.7510 ns |  1.02 |    0.20 |
| 'New Dictionary this[i] setter' | 4.651 ns | 0.1837 ns | 0.5181 ns |  0.92 |    0.16 |
```

Also:
* Pre-store `entries.Length` into variable to avoid extracting it two times on every iteration.

`CollectionsMarshalHelper.GetValueRefOrAddDefault()` is modified for consistency