@hez2010 2024-09-23
> avoid extracting it two times on every iteration

I don't think this should be done. We were intently calling `array.Length` in the branch condition every time because the JIT can recognize the pattern to remove unnecessary bound checks.