@SergeiPavlov 2024-09-23
> I don't think this should be done.

This JIT is not always smart.
I'd suggest to change floowing code:
```
uint collisionCount = 0;
...
collisionCount++;
if (collisionCount > (uint)entries.Length)
    throw ...;
```
to 
```
uint collisionsRemaining = entries.Length;
...
if (--collisionsRemaining == )
    throw ...;
```
Comaring with zero would be easier