@huoyaoyuan 2024-09-23
> Comparing with zero would be easier

JIT already does this if it can prove for safety.

The code you are touching includes some tricks for JIT. We should inspect the assembly code to ensure everything is working properly.