@SergeiPavlov 2024-09-23
> We should inspect the assembly code

Looked at the asm with `[DisassemblyDiagnoser]`:
```
   --collisionCount;
   if (collisionCount == 0)
```
translates into
```
       dec       r11d
       je        near ptr M01_L22
```

and the current code
```
if (collisionCount > (uint)entries.Length)
```
translates into
```
       inc       r12d
       cmp       [rsp+30],r12d
       jae       near ptr M01_L09
```

So JIT have proved that `entries.Length` can be saved in local var `[rsp+30]`, but it is still more CPU instructions