@huoyaoyuan 2024-09-23
> > We should inspect the assembly code
> 
> Looked at the asm with `[DisassemblyDiagnoser]`:

I'm referring to the disassembly for current change. Based on my knowledge I'd expect the change to result in zero difference.

> Comparing with zero would be easier

This is already an optimization employed by JIT:

https://devblogs.microsoft.com/dotnet/performance-improvements-in-net-9/#loops