@SergeiPavlov 2024-09-23
> I'm referring to the disassembly for current change. Based on my knowledge I'd expect the change to result in zero difference.

Well
Translation of
```
if (entries[i].hashCode == hashCode && EqualityComparer<TKey>.Default.Equals(entries[i].key, key))
```
to
```
       mov       r9d,r10d
       shl       r9,4
       cmp       [r14+r9+10],r13d
       jne       short M01_L08
       cmp       [r14+r9+18],esi
```

```
ref Entry e = ref entries[i];
if (e.hashCode == hashCode && EqualityComparer<TKey>.Default.Equals(e.key, key))
```
to
```
       mov       eax,r8d
       shl       rax,4
       lea       r8,[r14+rax+10]
       cmp       [r8],r13d
       jne       near ptr M01_L11
       cmp       [r8+8],esi
```

The code looks equivalently efficient.
But the final criterion is performance. I see ~10% improvement over many benchmarks runs