@eiriktsarpalis 2024-09-23
Have you tried running the full benchmark suite as defined in https://github.com/dotnet/performance/tree/main/src/benchmarks/micro/libraries/System.Collections?